fix(footer): update selector for even rows in contentArea table fix(cancel): correct type conversion for request ID in CancelMap chore: align html selectors between fields A, B, and C refactor(main): enhance TableStruct consistency by updating selector fix(selector): correct row odd to row even for TableStruct field C style(logging): add detailed log messages for request cancelation

+++ tools/seltabls/pkg/server/handle.go
@@ -127,7 +127,7 @@ func HandleMessage(
 					)
 				}
 				log.Debugf("canceling request: %d", request.Params.ID)
-				c, ok := lsp.CancelMap.Get(request.Params.ID.(int))
+				c, ok := lsp.CancelMap.Get(int(request.Params.ID.(float64)))
 				if !ok {
 					return nil, fmt.Errorf("failed to get cancel func")
 				}
